+++ packages/react/rollup.config.ts
@@ -4,37 +4,75 @@ import typescript from "rollup-plugin-typescript2";
 import postcss from "rollup-plugin-postcss";
 import { terser } from "rollup-plugin-terser";
 import pkg from "./package.json";
+import copy from "rollup-plugin-copy";
 
-export default {
-  input: "src/index.ts",
-  output: [
-    {
-      file: pkg.main,
-      format: "cjs",
-    },
-    {
-      file: pkg.module,
-      format: "es",
-    },
-  ],
-  external: ["react", "react-dom"],
-  plugins: [
-    resolve(),
-    commonjs(),
-    postcss({
-      extract: "styles.css",
-      config: {
-        path: "./postcss.config.js",
+export default [
+  {
+    input: "src/index.ts",
+    output: [
+      {
+        file: pkg.main,
+        format: "cjs",
       },
-      extensions: [".css"],
-      minimize: true,
-      inject: {
-        insertAt: "top",
+      {
+        file: pkg.module,
+        format: "es",
       },
-    }),
-    typescript({
-      useTsconfigDeclarationDir: true,
-    }),
-    // terser(),
-  ],
-};
+    ],
+    external: ["react", "react-dom"],
+    plugins: [
+      copy({
+        targets: [{ src: "swiper.css", dest: "dist/vendors" }],
+        copyOnce: true,
+      }),
+      resolve(),
+      commonjs(),
+      postcss({
+        extract: "styles.css",
+        config: {
+          path: "./postcss.config.js",
+          ctx: null,
+        },
+        extensions: [".css"],
+        minimize: true,
+        inject: {
+          insertAt: "top",
+        },
+      }),
+      typescript({
+        useTsconfigDeclarationDir: true,
+      }),
+      // terser(),
+    ],
+  },
+  // {
+  //   input: "src/swiper.css",
+  //   output: [
+  //     {
+  //       file: pkg.main,
+  //       format: "cjs",
+  //     },
+  //     {
+  //       file: pkg.module,
+  //       format: "es",
+  //     },
+  //   ],
+  //   external: ["react", "react-dom"],
+  //   plugins: [
+  //     resolve(),
+  //     commonjs(),
+  //     postcss({
+  //       extract: "swiper.css",
+  //       config: {
+  //         path: "./postcss.config.js",
+  //         ctx: null,
+  //       },
+  //       extensions: [".css"],
+  //       minimize: true,
+  //       // inject: {
+  //       //   insertAt: "top",
+  //       // },
+  //     }),
+  //   ],
+  // }
+];