+++ packages/react/src/molecules/modals/confirm.tsx
@@ -50,11 +50,21 @@ const ModalConfirm = (props: ModalConfirmProps) => {
   );
 };
 
-export const canBeClosed = mountComponent(Closable(ModalConfirm), 500);
+// export const canBeClosed = mountComponent(Closable(ModalConfirm), 500);
 
-function Exportable(config: any, options = {}) {
-  config.timeout = 0;
-  return canBeClosed({ ...config, ...options });
+function Exportable(
+  config: ModalConfirmProps,
+  unmountDelay = 1000,
+  mountingNode?: any
+) {
+  if (config.timeout === undefined) {
+    config.timeout = 0;
+  }
+  return mountComponent(
+    Closable(ModalConfirm),
+    unmountDelay,
+    mountingNode
+  )(config);
 }
 
 export default Exportable;